Use a float for new tile spawn; fixes #60

+++ game_shared.c
@@ -113,7 +113,7 @@ static void add_tile(void)
       empty[j]->source = NULL;
       empty[j]->move_time = 1;
       empty[j]->appear_time = 0;
-      empty[j]->value = (rand() / RAND_MAX) < 0.9 ? 1 : 2;
+      empty[j]->value = ((float)rand() / RAND_MAX) < 0.9 ? 1 : 2;
    }
    else
       change_state(STATE_GAME_OVER);
